Rename all Stm32f1 to STM32F1 (#12639)
