+++ Marlin/src/HAL/HAL_STM32F1/HAL.h
@@ -47,10 +47,10 @@
 #include "../shared/math_32bit.h"
 #include "../shared/HAL_SPI.h"
 
-#include "fastio_Stm32f1.h"
-#include "watchdog_Stm32f1.h"
+#include "fastio_STM32F1.h"
+#include "watchdog_STM32F1.h"
 
-#include "HAL_timers_Stm32f1.h"
+#include "HAL_timers_STM32F1.h"
 
 
 // --------------------------------------------------------------------------
