+++ Marlin/src/HAL/HAL_STM32F1/HAL_Servo_STM32F1.cpp
@@ -29,7 +29,7 @@
 
 uint8_t ServoCount; //=0
 
-#include "HAL_Servo_Stm32f1.h"
+#include "HAL_Servo_STM32F1.h"
 
 //#include "Servo.h"
