+++ Marlin/src/HAL/HAL_STM32F1/fastio_STM32F1.h
