+++ platformio.ini
@@ -252,7 +252,7 @@ monitor_speed = 250000
 platform      = ststm32@<4.4.0
 framework     = arduino
 board         = genericSTM32F103RE
-build_flags   = !python Marlin/src/HAL/HAL_STM32F1/stm32f1_flag_script.py
+build_flags   = !python Marlin/src/HAL/HAL_STM32F1/STM32F1_flag_script.py
   ${common.build_flags}
 lib_deps      = ${common.lib_deps}
 lib_ignore    = U8glib-HAL
@@ -301,7 +301,7 @@ platform      = ststm32
 framework     = arduino
 board         = genericSTM32F103ZE
 extra_scripts = buildroot/share/PlatformIO/scripts/mks_robin.py
-build_flags   = !python Marlin/src/HAL/HAL_STM32F1/stm32f1_flag_script.py
+build_flags   = !python Marlin/src/HAL/HAL_STM32F1/STM32F1_flag_script.py
   ${common.build_flags}
 src_filter    = ${common.default_src_filter} +<src/HAL/HAL_STM32F1>
 lib_deps      = ${common.lib_deps}
@@ -330,7 +330,7 @@ monitor_speed = 250000
 platform    = ststm32
 framework   = arduino
 board       = malyanM200
-build_flags = !python Marlin/src/HAL/HAL_STM32F1/stm32f1_flag_script.py -DMCU_STM32F103CB -D __STM32F1__=1 -std=c++1y -D MOTHERBOARD="BOARD_MALYAN_M200" -DSERIAL_USB -ffunction-sections -fdata-sections -Wl,--gc-sections
+build_flags = !python Marlin/src/HAL/HAL_STM32F1/STM32F1_flag_script.py -DMCU_STM32F103CB -D __STM32F1__=1 -std=c++1y -D MOTHERBOARD="BOARD_MALYAN_M200" -DSERIAL_USB -ffunction-sections -fdata-sections -Wl,--gc-sections
 src_filter  = ${common.default_src_filter} +<src/HAL/HAL_STM32F1>
 #-<frameworks>
 lib_ignore  =
